♻️ refactor redirection at login in dedicated hook
